Clean up the crate structure a bit
