+++ native/ssz_nif/src/utils/helpers.rs
@@ -1,0 +1,32 @@
+use crate::utils::from_lh::FromLH;
+use rustler::{Binary, Decoder, Encoder, Env, NewBinary, NifResult, Term};
+use ssz::{Decode, Encode};
+use std::io::Write;
+
+pub(crate) fn bytes_to_binary<'env>(env: Env<'env>, bytes: &[u8]) -> Binary<'env> {
+    let mut binary = NewBinary::new(env, bytes.len());
+    // This cannot fail because bin size equals bytes len
+    binary.as_mut_slice().write_all(bytes).unwrap();
+    binary.into()
+}
+
+pub(crate) fn encode_ssz<'a, Elx, Lh>(value: Term<'a>) -> NifResult<Vec<u8>>
+where
+    Elx: Decoder<'a> + Into<Lh>,
+    Lh: Encode,
+{
+    let value_nif = <Elx as Decoder>::decode(value)?;
+    let value_ssz: Lh = value_nif.into();
+    Ok(value_ssz.as_ssz_bytes())
+}
+
+pub(crate) fn decode_ssz<'a, Elx, Lh>(bytes: &[u8], env: Env<'a>) -> NifResult<Term<'a>>
+where
+    Elx: Encoder + FromLH<'a, Lh>,
+    Lh: Decode,
+{
+    let recovered_value = Lh::from_ssz_bytes(bytes).expect("can deserialize");
+    let checkpoint = Elx::from(recovered_value, env);
+    let term = checkpoint.encode(env);
+    Ok(term)
+}
